Now .hook.yml based with multi-level triggering
The functions will now scan for a file called .hooks.yml in the originating repo and act based on the hooks configured therein. This has reduced the call strucutre to a single ingestion point and a "switch-like" secondary call for desired secondary functionality.

Also adding tons of safety with more checking for validation, yml schema validations, and better logging.